last

+++ Triangle.java
@@ -4,15 +4,17 @@
  * Write a program that tests if three given integers form a triangle.
  */ 
 public class Triangle {
-	public static void main(String[] args) {
-		double a = Double.parseDouble(args[0]);
-    double b = Double.parseDouble(args[1]);
-    double c = Double.parseDouble(args[2]);
-    if ((a+b<c)|| (a+c<b) || (b+c<a)){
+    public static void main(String[] args) {
+    int a = Integer.parseInt(args[0]);
+    int b = Integer.parseInt(args[1]);
+    int c = Integer.parseInt(args[2]);
+   
+    if ((a+b<c) || (a+c<b) || (b+c<a)){
         System.out.println(a+", "+ b+", "+c+": "+"false");
     }
     else {
         System.out.println(a+", "+ b+", "+c+": "+"true");
     } 
-	}
+
+    }
 }
